move transposer_engine_internal to its own file
