+++ src/core/transposer/mod.rs
@@ -3,3 +3,4 @@ pub mod transposer;
 pub mod transposer_engine;
 pub mod transposer_event;
 pub mod transposer_context;
+pub(crate) mod transposer_engine_internal;
